+++ react/__tests__/navigation.test.js
@@ -28,6 +28,17 @@ describe('Navigation route modifier', () => {
 
     expect(result).toEqual({ path: expectedPath, query: expectedQuery })
   })
+  it('intelligent search queries should have the following order: /[category hierarchy/alphabetically]/[brand alphabetically]/[specs alphabetically]', () => {
+    const path = '/departmentvalue/categoryvalue1/categoryvalue2'
+    const query =
+      'map=specificationFilter,category-2,category-2,brand,category-1,brand,&query=/specvalue/categoryvalue2/categoryvalue1/brandvalue2/departmentvalue/brandvalue1'
+    const expectedPath = path.toLocaleLowerCase()
+    const expectedQuery =
+      'map=category-1,category-2,category-2,brand,brand,specificationFilter&query=/departmentvalue/categoryvalue1/categoryvalue2/brandvalue1/brandvalue2/specvalue'
+    const result = normalizeNavigation({ path, query })
+
+    expect(result).toEqual({ path: expectedPath, query: expectedQuery })
+  })
   it('should preserve case for new routes pattern', () => {
     const path = '/foo/bar/nice_Bar'
     const expectedPath = '/foo/bar/nice_Bar'