+++ react/utils/navigation.js
@@ -7,6 +7,7 @@ const SPEC_FILTER = 'specificationFilter'
 const MAP_CATEGORY_CHAR = 'c'
 const MAP_VALUES_SEP = ','
 const PATH_SEPARATOR = '/'
+const INTELLIGENT_SEARCH_BRAND = 'brand'
 
 const isLegacySearchFormat = (pathSegments, map) => {
   if (!map) {
@@ -29,8 +30,36 @@ const categoriesInSequence = mapValues => {
 const queryMapComparator = (tuple1, tuple2) => {
   const [, specFilterVal1] = tuple1 && tuple1[0].split(`${SPEC_FILTER}_`)
   const [, specFilterVal2] = tuple2 && tuple2[0].split(`${SPEC_FILTER}_`)
-  const facetName1 = tuple1[1]
-  const facetName2 = tuple2[1]
+  const [facetName1, facetValue1] = tuple1
+  const [facetName2, facetValue2] = tuple2
+
+  const intelligentSearchCategoryRegex = /category-[0-9]+/
+  const isISCategory1 = intelligentSearchCategoryRegex.test(facetName1)
+  const isISCategory2 = intelligentSearchCategoryRegex.test(facetName2)
+  const isISBrand1 = facetName1 === INTELLIGENT_SEARCH_BRAND
+  const isISBrand2 = facetName2 === INTELLIGENT_SEARCH_BRAND
+
+  // Makes the URL in the IS have the following order:
+  // /[category hierarchy/alphabetically]/[brand alphabetically]/[specs alphabetically]
+  // Example: category-1/category-2/brand/spec1/spec2
+  if (isISCategory1 && !isISCategory2) {
+    return -1
+  }
+  if (isISCategory2 && !isISCategory1) {
+    return 1
+  }
+  if (isISCategory1 && isISCategory2) {
+    return (
+      facetName1.localeCompare(facetName2) ||
+      facetValue1.localeCompare(facetValue2)
+    )
+  }
+  if (isISBrand1 && !isISBrand2) {
+    return -1
+  }
+  if (isISBrand2 && !isISBrand1) {
+    return 1
+  }
 
   const considerSpecificationFilter =
     specFilterVal1 &&
@@ -41,8 +70,8 @@ const queryMapComparator = (tuple1, tuple2) => {
   return considerSpecificationFilter
     ? Number(specFilterVal1) -
         Number(specFilterVal2) +
-        facetName1.localeCompare(facetName2)
-    : facetName1.localeCompare(facetName2)
+        facetValue1.localeCompare(facetValue2)
+    : facetValue1.localeCompare(facetValue2)
 }
 
 const normalizeQueryMap = (pathSegments, mapSegments) => {